bakcup

+++ gaze/input_utils.py
@@ -23,8 +23,8 @@ class DatasetWithHeatmap(BIU.Dataset):
     super(DatasetWithHeatmap, self).__init__(LABELS_FILE_TRAIN, LABELS_FILE_VAL, RESIZE_SHAPE)
     print "Reading gaze data ASC file, and converting per-frame gaze positions to heat map..."
     self.frameid2pos, self.frameid2action_notused, _, _, _ = BIU.read_gaze_data_asc_file(GAZE_POS_ASC_FILE)
-    self.train_GHmap = np.zeros([self.train_size, HEATMAP_SHAPE, HEATMAP_SHAPE, 1], dtype=np.float32)
-    self.val_GHmap = np.zeros([self.val_size, HEATMAP_SHAPE, HEATMAP_SHAPE, 1], dtype=np.float32)
+    self.train_GHmap = np.zeros([self.train_size, HEATMAP_SHAPE, HEATMAP_SHAPE, 1], dtype=np.float16)
+    self.val_GHmap = np.zeros([self.val_size, HEATMAP_SHAPE, HEATMAP_SHAPE, 1], dtype=np.float16)
 
     # Prepare train val gaze data
     print "Running BIU.convert_gaze_pos_to_heap_map() and convolution..."
@@ -43,8 +43,8 @@ class DatasetWithHeatmap(BIU.Dataset):
 
     sigmaH = 28.50 * HEATMAP_SHAPE / V.SCR_H
     sigmaW = 44.58 * HEATMAP_SHAPE / V.SCR_W
-    self.train_GHmap = BIU.preprocess_gaze_heatmap(self.train_GHmap, sigmaH, sigmaW, 0).astype(np.float32)
-    self.val_GHmap = BIU.preprocess_gaze_heatmap(self.val_GHmap, sigmaH, sigmaW, 0).astype(np.float32)
+    self.train_GHmap = BIU.preprocess_gaze_heatmap(self.train_GHmap, sigmaH, sigmaW, 0).astype(np.float16)
+    self.val_GHmap = BIU.preprocess_gaze_heatmap(self.val_GHmap, sigmaH, sigmaW, 0).astype(np.float16)
 
     print "Normalizing the train/val heat map..."
     for i in range(len(self.train_GHmap)):
@@ -167,7 +167,7 @@ def read_optical_flow(label_file, RESIZE_SHAPE, num_thread=6):
             png_files.append(fname)
 
     N = len(png_files)
-    imgs = np.empty((N,RESIZE_SHAPE[0],RESIZE_SHAPE[1],1), dtype=np.float32)
+    imgs = np.empty((N,RESIZE_SHAPE[0],RESIZE_SHAPE[1],1), dtype=np.float16)
 
     def read_thread(PID):
         d = os.path.dirname(label_file)
@@ -175,10 +175,10 @@ def read_optical_flow(label_file, RESIZE_SHAPE, num_thread=6):
             try:
                 img = misc.imread(os.path.join(d+'/optical_flow', png_files[i]))
             except IOError:
-                img = np.zeros((RESIZE_SHAPE[0],RESIZE_SHAPE[1]), dtype=np.float32)
+                img = np.zeros((RESIZE_SHAPE[0],RESIZE_SHAPE[1]), dtype=np.float16)
                 print "Warning: %s has no optical flow image. Set to zero." % png_files[i]
             img = np.expand_dims(img, axis=2)
-            img = img.astype(np.float32) / 255.0 # normalize image to [0,1]            
+            img = img.astype(np.float16) / 255.0 # normalize image to [0,1]            
             imgs[i,:] = img
 
     o=BIU.ForkJoiner(num_thread=num_thread, target=read_thread)
@@ -197,7 +197,7 @@ def read_bottom_up(label_file, RESIZE_SHAPE, num_thread=6):
             png_files.append(fname)
 
     N = len(png_files)
-    imgs = np.empty((N,RESIZE_SHAPE[0],RESIZE_SHAPE[1],1), dtype=np.float32)
+    imgs = np.empty((N,RESIZE_SHAPE[0],RESIZE_SHAPE[1],1), dtype=np.float16)
 
     def read_thread(PID):
         d = os.path.dirname(label_file)
@@ -205,10 +205,10 @@ def read_bottom_up(label_file, RESIZE_SHAPE, num_thread=6):
             try:
                 img = misc.imread(os.path.join(d+'/bottom_up', png_files[i]))
             except IOError:
-                img = np.zeros((RESIZE_SHAPE[0],RESIZE_SHAPE[1]), dtype=np.float32)
+                img = np.zeros((RESIZE_SHAPE[0],RESIZE_SHAPE[1]), dtype=np.float16)
                 print "Warning: %s has no bottom up image. Set to zero." % png_files[i]
             img = np.expand_dims(img, axis=2)
-            img = img.astype(np.float32) / 255.0 # normalize image to [0,1]            
+            img = img.astype(np.float16) / 255.0 # normalize image to [0,1]            
             imgs[i,:] = img
 
     o=BIU.ForkJoiner(num_thread=num_thread, target=read_thread)
