+++ shared/base_input_utils.py
@@ -196,8 +196,8 @@ def read_np_parallel(label_file, RESIZE_SHAPE, num_thread=6, preprocess_deprecat
     N = len(labels)
     imgs = [None] * N
     labels = np.asarray(labels, dtype=np.int32)
-    gaze = np.asarray(gaze, dtype=np.float32)
-    weight = np.asarray(weight, dtype=np.float32)
+    gaze = np.asarray(gaze, dtype=np.float16)
+    weight = np.asarray(weight, dtype=np.float16)
 
     def read_thread(PID):
         d = os.path.dirname(label_file)
@@ -206,7 +206,7 @@ def read_np_parallel(label_file, RESIZE_SHAPE, num_thread=6, preprocess_deprecat
                 img = misc.imread(os.path.join(d, png_files[i]), 'Y') # 'Y': grayscale  
                 img = misc.imresize(img, [RESIZE_SHAPE[0],RESIZE_SHAPE[1]], interp='bilinear')
                 img = np.expand_dims(img, axis=2)
-                img = img.astype(np.float32) / 255.0 # normalize image to [0,1]
+                img = img.astype(np.float16) / 255.0 # normalize image to [0,1]
             else:
                 img = misc.imread(os.path.join(d, png_files[i])) # uint8 RGB (210,160,3)
             imgs[i] = img
